修复录制当大小超过int最大值分割失败的问题

+++ src/main/java/moe/chikalar/bili/dto/RecordConfig.java
@@ -19,7 +19,7 @@ public class RecordConfig {
     private Boolean splitFileByTime = false;
     private Integer splitFileDurationInSeconds = 3600;
     private Boolean splitFileBySize = false;
-    private Integer splitFileSizeInM = 4000;
+    private Long splitFileSizeInM = 4000L;
     private Boolean splitFileDeleteSource = false;
